Built initial unit test and test scripts
Developed a unit test to test the current core functionality. The unit
test consists of a main app that calls the tests and a suite of tests.
The unit testing reports are very raw. Also in this commit are the test
scripts that have the test dates, and a sample set of holidy rules.