Add new set_weekday function
With the existing api, a user must set the day_of_week member of a
DateTime struct, by assigning the return value of derive_weekday to the
very struct used as an argument to the function call. With set_weekday,
a user can just directly set the weekday through the function call and
avoid making an explicit assignment.