Removed test_loop in favor of testsuite_run_check
The functions testsuite_check_calcs, testsuite_check_deriveday,
testsuite_check_holidays, and testsuite_check_leap only served to open a
file, print a few messages, call the loop, and close the file. All the
logic for each test was in the loop itself. This would make sense if the
the testing was the same and we were just looping. However, the logic
was in each case statement and it differed for each case.

It made more sense to reverse the process. Make one function that calls
the appropriate check routine. This seems like it could still be
improved with refactoring to eliminate some redundancy, but the logic is
more clear and the readability much improved.