Added testsuite_interactive
Added an interactive mode, and a script to launch it. So far it only
tests one date and then ends. I moved the other test batteries into the
commandline processing switch statement. The only trouble to note is
that the holiday file must be the first commandline argument otherwise
the test will fail.