Debugged courtday_difference; added testsuite
Fixed the "FIXME" where courtday_difference would not correctly compute
the number of courtdays if the starting date began on a holiday. Also
changed the function to pass the date structs by value rather than
through pointers so that the arguments could not accidentally be
changed.

Added a testsuite to test the courtday_difference function. It worked
properly on all dates tested thus far. Note that because not actual
court rules file has been built yet, court days only exclude weekeneds
and other holidays in the holiday file.