Fixed file handling to deal with CRLF line endings
The library handles any type of line ending for the test files through
the standard C library functions that read from them. The function
holiday_rules_tokenize, however, uses its own parsing routine
that previously did not handle CRLF endings. It was fixed simply by
defining CARRIAGE_RTN along with the other field delimiters, and then
adding a fall-through case for it where NEWLINE is handled in the
*cur_char switch statement. In the case for TDELIMITER, CARRIAGE_RTN is
also referenced and is treated as being equivalent to NEWLINE. Since the
next iteration in the loop calling this function simply reads a newline
from the input file, we never have to actually consume any NEWLINE that
might come after the CARRIAGE_RTN.

Added new rules and testscripts for testing the CRLF handling. Same
rules and tests as the existing test battery, but with CRLF endings.

Updated the .gitattributes file to allow those CRLF rules and test
scripts to be commited  wit their CRLF endings intact.

Added the test script for courtday_difference and courtday_offset test suites.