Added the caldays_test.csv test script
Neglected to add the caldays_test test-script in commit ec92d00.
Updated .gitignore to exclude my playground directory (used as a temp
directory for stashing experimental files that should not be tracked).
Also temporarily ignored a few doc directories because those
documentation files are not ready for their first commit.

Also note: commit a8d2179 was a accident and should be ignored.