Addressed two TODO comments. Added a #define for weekend rules.
In adding error handling to the library, the first step is addressing the open TODOS. No change was made to the TODO regarding having jdncnvrt return an int, but simply assign the JDN to the struct member data object jdn.
After reviewing the code, it was, in a number of places, the return value is used to perform a computation. It seemed preferable to leave it as is. The other TODO that was addressed was to create a symbolic constant for
weekend rules. A weekend rule sets the day of the month to zero. That is now #defined as WKNDRULE_DAY.